Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,10 +2,10 @@
 int main() {
     printf("");
     int a,b,c,d;
-    double e;
+    float e;
     scanf("%d &d %d",&a,&b,&c);
     d=a+b+c;
     e=d/3;
-    printf("Average: %.2lf",e);
+    printf("Average: %.2f",e);
     return 0;
 }
